refactor loyalty rule update validation and integer discounts

+++ heatmap-api/src/main/java/ru/tbank/tmap/loyalty/business/BusinessLoyaltyRuleCreateCommand.java
@@ -1,10 +1,8 @@
 package ru.tbank.tmap.loyalty.business;
 
-import java.math.BigDecimal;
-
 public record BusinessLoyaltyRuleCreateCommand(
         String description,
-        BigDecimal discountPercent,
+        int discountPercent,
         int maxUsages
 ) {
 }
